refactor(Obra): Refatoração controller Obra
