+++ TsundokuTraducoes.Helpers/Validacao/ValidacaoRequest.cs
@@ -1,9 +1,59 @@
-﻿using TsundokuTraducoes.Helpers.DTOs.Public.Request;
+﻿using System.Text.RegularExpressions;
+using TsundokuTraducoes.Helpers.DTOs.Admin;
+using TsundokuTraducoes.Helpers.DTOs.Public.Request;
 
 namespace TsundokuTraducoes.Helpers.Validacao
 {
     public static class ValidacaoRequest
     {
+        public static bool ValidaDadosRequestObra(ObraDTO obraDTO)
+        {
+            var resquestValido = VerificaString(obraDTO.Titulo) &&
+                VerificaString(obraDTO.Alias) &&
+                VerificaString(obraDTO.TituloAlternativo) &&
+                VerificaString(obraDTO.Autor) &&
+                VerificaString(obraDTO.Ano) &&
+                VerificaString(obraDTO.UsuarioInclusao) &&
+                VerificaString(obraDTO.Sinopse) &&
+                VerificaString(obraDTO.CodigoCorHexaObra) &&
+                VerificaString(obraDTO.NacionalidadeSlug) &&
+                VerificaString(obraDTO.StatusObraSlug) &&
+                VerificaString(obraDTO.TipoObraSlug) &&
+                obraDTO.ListaGeneros.Count > 0 &&
+                obraDTO.ImagemCapaPrincipalFile != null;
+
+            return resquestValido;
+        }
+
+        public static bool ValidaDadosRequestObraAtualizacao(ObraDTO obraDTO)
+        {
+            var resquestValido = VerificaString(obraDTO.Titulo) &&
+                VerificaString(obraDTO.Alias) &&
+                VerificaString(obraDTO.TituloAlternativo) &&
+                VerificaString(obraDTO.Autor) &&
+                VerificaString(obraDTO.Ano) &&
+                VerificaString(obraDTO.UsuarioInclusao) &&
+                VerificaString(obraDTO.Sinopse) &&
+                VerificaString(obraDTO.CodigoCorHexaObra) &&
+                VerificaString(obraDTO.NacionalidadeSlug) &&
+                VerificaString(obraDTO.StatusObraSlug) &&
+                VerificaString(obraDTO.TipoObraSlug) &&
+                obraDTO.ListaGeneros.Count > 0;
+
+            return resquestValido;
+        }
+
+        public static bool VerificaString(string valor)
+        {
+            return !string.IsNullOrEmpty(valor) && !valor.Contains("\"\""); ;
+        }
+
+        public static bool ValidaCorHexaDecimal(string corHexaDeximal)
+        {
+            var regexPattern = "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$";
+            return Regex.Match(corHexaDeximal, regexPattern).Success;
+        }
+
         public static bool ValidaParametrosNovel(RequestObras requestObras)
         {
             return ValidaParametrosObra(requestObras);
